Added Readme.md

+++ RobotWars/RobotWars/Robot.cs
@@ -7,6 +7,6 @@
         public int Attack;
         public int Defense;
         public int HP;
-		public decimal Cost;
+		public double Cost;
     }
 }
